requirements.txt
playwright==1.40.0
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.1.4
sqlalchemy==2.0.23
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.3.7
transformers==4.35.2
torch==2.1.1
selenium==4.15.2
lxml==4.9.3
fake-useragent==1.4.0
python-telegram-bot==20.7
schedule==1.2.0
